Move txqueue code into txqueue package
