+++ geth/txqueue/txqueue.go
@@ -1,4 +1,4 @@
-package node
+package txqueue
 
 import (
 	"errors"
