+++ geth/txqueue/txqueue_manager.go
@@ -1,4 +1,4 @@
-package node
+package txqueue
 
 import (
 	"context"
