+++ geth/txqueue/txqueue_manager_test.go
@@ -1,4 +1,4 @@
-package node
+package txqueue
 
 import (
 	"context"
